# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ZKshare API",
    "version": "1.0.0",
    "description": "ZKshare — MCP-native + REST. Operations: store (server-sealed or client-sealed ciphertext bundle), prove, share, search, verify_proof, sandbox. Tiers: Free 1k ops/mo, Starter $19 / 20k, Pro $49 / 100k + audit export, Enterprise custom. Metered overages via Stripe ($0.005–$0.02/op typical)."
  },
  "servers": [{ "url": "/", "description": "Same origin as Next.js app" }],
  "paths": {
    "/api/v1/context": {
      "post": {
        "summary": "Context operations",
        "security": [{ "ApiKeyAuth": [] }],
        "parameters": [
          {
            "name": "x-request-id",
            "in": "header",
            "required": false,
            "schema": { "type": "string", "format": "uuid" },
            "description": "Optional idempotency / tracing id (echoed on response)"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ContextRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ContextSuccess" }
              }
            }
          },
          "400": {
            "description": "Validation or proof failure",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ContextError" }
              }
            }
          },
          "401": {
            "description": "Invalid or revoked API key",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ContextError" }
              }
            }
          },
          "404": {
            "description": "Fact not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ContextError" }
              }
            }
          },
          "413": {
            "description": "Payload too large",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ContextError" }
              }
            }
          },
          "422": {
            "description": "Client-sealed fact cannot be processed server-side (e.g. prove/share)",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ContextError" }
              }
            }
          },
          "429": {
            "description": "Rate limited (includes Retry-After header)",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ContextError" }
              }
            }
          }
        }
      }
    },
    "/api/health": {
      "get": {
        "summary": "Liveness (edge; no database)",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": { "type": "boolean" },
                    "service": { "type": "string" },
                    "timestamp": { "type": "string", "format": "date-time" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/health/ready": {
      "get": {
        "summary": "Readiness (database connectivity)",
        "responses": {
          "200": {
            "description": "Database reachable",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": { "type": "boolean" },
                    "database": { "type": "boolean" },
                    "timestamp": { "type": "string", "format": "date-time" }
                  }
                }
              }
            }
          },
          "503": {
            "description": "Not ready",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": { "type": "boolean" },
                    "database": { "type": "boolean" },
                    "reason": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    },
    "schemas": {
      "ContextRequest": {
        "type": "object",
        "required": ["operation"],
        "properties": {
          "operation": {
            "type": "string",
            "enum": ["store", "prove", "share", "search", "sandbox", "verify_proof"]
          },
          "user_id": { "type": "string", "description": "Logical user scope (recommended)" },
          "fact_key": { "type": "string" },
          "value": {
            "type": "string",
            "description": "For store: plaintext when using server-side sealing (mutually exclusive with ciphertext bundle)"
          },
          "ciphertext": {
            "type": "string",
            "description": "For store (client-sealed): AES-GCM ciphertext (base64 or opaque string as stored)"
          },
          "iv": { "type": "string", "description": "For store (client-sealed): AES-GCM IV/nonce" },
          "auth_tag": { "type": "string", "description": "For store (client-sealed): AES-GCM auth tag" },
          "commitment": {
            "type": "string",
            "description": "For store (client-sealed): commitment string (same scheme as server-sealed facts)"
          },
          "proof": {
            "type": "string",
            "description": "For verify_proof: base64url proof string from a prior prove or share response"
          },
          "query": { "type": "string" },
          "recipient_agent_id": { "type": "string" },
          "action": { "type": "string" },
          "parameters": { "type": "object", "additionalProperties": true },
          "embedding": {
            "type": "array",
            "items": { "type": "number" },
            "minItems": 1536,
            "maxItems": 1536,
            "description": "Required for client-sealed store (server must not derive embeddings). Optional for server-sealed store (skips server embed call when set). Use the same 1536-dim model/projection as the platform embedding pipeline."
          }
        }
      },
      "ContextSuccess": {
        "type": "object",
        "required": ["success", "operation", "data", "proof", "verified", "timestamp", "usage"],
        "properties": {
          "success": { "type": "boolean", "const": true },
          "operation": { "type": "string" },
          "data": { "type": "object", "additionalProperties": true },
          "proof": { "type": ["string", "null"] },
          "verified": { "type": "boolean" },
          "timestamp": { "type": "string", "format": "date-time" },
          "request_id": { "type": "string", "description": "Correlation id for logs / support" },
          "usage": {
            "type": "object",
            "properties": {
              "calls": { "type": "integer" },
              "limit": { "type": "integer" }
            }
          }
        }
      },
      "ContextError": {
        "type": "object",
        "required": ["success", "error", "message"],
        "properties": {
          "success": { "type": "boolean", "const": false },
          "request_id": { "type": "string" },
          "error": {
            "type": "string",
            "enum": [
              "INVALID_API_KEY",
              "RATE_LIMITED",
              "VALIDATION_ERROR",
              "FACT_NOT_FOUND",
              "PROOF_FAILED",
              "INTERNAL_ERROR",
              "CONFIG_ERROR",
              "CLIENT_ENCRYPTED"
            ]
          },
          "message": { "type": "string" }
        }
      }
    }
  }
}